META-237: Fixed compilation warning
<==== Mercurial data ====>
Branch: vms_4.0_server

+++ vms/libs/common/src/analytics/detected_objects_storage/object_searcher.cpp
@@ -236,7 +236,7 @@ void ObjectSearcher::filterTrack(
         track->begin(), track->end(),
         [](const auto& left, const auto& right) { return left.timestampUsec < right.timestampUsec; });
 
-    if (filter.maxTrackSize > 0 && track->size() > filter.maxTrackSize)
+    if (filter.maxTrackSize > 0 && (int) track->size() > filter.maxTrackSize)
         track->erase(track->begin() + filter.maxTrackSize, track->end());
 }
